+++ src/cursorless.talon
@@ -1,8 +1,8 @@
 app: vscode
 -
 
-{user.cursorless_simple_action} <user.cursorless_target>:
-    user.cursorless_simple_action(cursorless_simple_action, cursorless_target)
+<user.cursorless_action> <user.cursorless_target>:
+    user.cursorless_action(cursorless_action, cursorless_target)
 
 {user.cursorless_swap_action} <user.cursorless_swap_targets>:
     user.cursorless_multiple_target_command(cursorless_swap_action, cursorless_swap_targets)
